+++ src/cornice_swagger/converters/parameters.py
@@ -13,7 +13,8 @@ class ParameterConverter(object):
 
         converted = {"name": schema_node.name, "in": self._in, "required": schema_node.required}
         if schema_node.description:
-            converted["description"] = schema_node.description
+            # keep 'description' for back-compatibility
+            converted['summary'] = converted['description'] = schema_node.description
 
         if schema_node.default:
             converted["default"] = schema_node.default
@@ -57,11 +58,24 @@ class BodyParameterConverter(ParameterConverter):
     def convert(self, schema_node, definition_handler):
         converted = {"name": schema_node.name, "in": self._in, "required": schema_node.required}
         if schema_node.description:
-            converted["description"] = schema_node.description
+            # keep 'description' for back-compatibility
+            converted['summary'] = converted['description'] = schema_node.description
 
         schema_node.title = schema_node.__class__.__name__
         schema = definition_handler(schema_node)
-        converted["schema"] = schema
+        converted['schema'] = schema
+
+        examples = getattr(schema_node, 'examples', None)
+        if examples and isinstance(examples, dict) and (
+                all(isinstance(name, str) and isinstance(ex, dict) for name, ex in examples.items())
+        ):
+            converted['examples'] = examples
+
+        examples = getattr(schema_node, 'examples', None)
+        if examples and isinstance(examples, dict) and (
+                all(isinstance(name, str) and isinstance(ex, dict) for name, ex in examples.items())
+        ):
+            converted['examples'] = examples
 
         return converted