+++ src/cornice_swagger/swagger.py
@@ -1,7 +1,9 @@
-"""Cornice Swagger 2.0 documentor"""
+"""Cornice Swagger 2.0 and OpenAPI 3 documentor"""
 import inspect
 import warnings
 from collections import OrderedDict
+from distutils.version import LooseVersion
+from typing import TYPE_CHECKING
 
 import colander
 from cornice import Service
@@ -11,6 +13,72 @@ from pyramid.threadlocal import get_current_registry
 from cornice_swagger.converters import ParameterConversionDispatcher as ParameterConverter
 from cornice_swagger.converters import TypeConversionDispatcher as TypeConverter
 from cornice_swagger.util import body_schema_transformer, merge_dicts, trim
+from cornice_swagger.converters import (TypeConversionDispatcher as TypeConverter,
+                                        ParameterConversionDispatcher as ParameterConverter)
+
+if TYPE_CHECKING:
+    import typing
+    from typing import Dict, List, Optional, Union
+    if hasattr(typing, "TypedDict"):
+        from typing import TypedDict
+    else:
+        from typing_extensions import TypedDict  # noqa
+
+    SwaggerInfo = TypedDict("SwaggerInfo", {"title": Optional[str], "version": Optional[str]}, total=False)
+    SwaggerTag = TypedDict("SwaggerTag", {"name": str}, total=False)
+    SwaggerDoc = TypedDict("SwaggerDoc", {"description": str, "url": str})
+    SwaggerRef = TypedDict("SwaggerRef", {"$ref": str})
+    SwaggerSummary = Optional[str]
+    SwaggerSchema = TypedDict("SwaggerSchema", {"type": str, "title": str,
+                                                "required": Optional[List[str]],
+                                                "items": Optional[List["SwaggerSchemaRef"]],
+                                                "properties": Dict[str, "SwaggerSchemaRef"]}, total=False)
+    SwaggerSchemaRef = Union[SwaggerSchema, SwaggerRef]
+    SwaggerContent = TypedDict("SwaggerContent", {"schema": SwaggerSchemaRef})
+    SwaggerResponse = TypedDict("SwaggerResponse", {"summary": SwaggerSummary, "content": Dict[str, SwaggerContent]})
+    SwaggerParameter = TypedDict("SwaggerParameter", {"name": str, "in": str, "required": bool, "type": str})
+    SwaggerSecurity = TypedDict("SwaggerSecurity", {"type": str, "name": str, "in": str})
+    SwaggerPath = TypedDict("SwaggerPath", {"responses": Dict[str, SwaggerResponse],
+                                            "parameters": List[SwaggerParameter],
+                                            "summary": SwaggerSummary,
+                                            "tags": List[str]}, total=False)
+    SwaggerType = TypedDict("SwaggerType", {"info": SwaggerInfo, "host": Optional[str], "schemes": List[str],
+                                            "basePath": str, "swagger": Optional[str], "openapi": Optional[str],
+                                            "tags": List[SwaggerTag], "paths": Dict[str, SwaggerPath],
+                                            "externalDocs": Optional[SwaggerDoc],
+                                            "securityDefinitions": Optional[Dict[str, SwaggerSecurity]]}, total=False)
+
+if TYPE_CHECKING:
+    import typing
+    from typing import Dict, List, Optional, Union
+    if hasattr(typing, "TypedDict"):
+        from typing import TypedDict
+    else:
+        from typing_extensions import TypedDict  # noqa
+
+    SwaggerInfo = TypedDict("SwaggerInfo", {"title": Optional[str], "version": Optional[str]}, total=False)
+    SwaggerTag = TypedDict("SwaggerTag", {"name": str}, total=False)
+    SwaggerDoc = TypedDict("SwaggerDoc", {"description": str, "url": str})
+    SwaggerRef = TypedDict("SwaggerRef", {"$ref": str})
+    SwaggerSummary = Optional[str]
+    SwaggerSchema = TypedDict("SwaggerSchema", {"type": str, "title": str,
+                                                "required": Optional[List[str]],
+                                                "items": Optional[List["SwaggerSchemaRef"]],
+                                                "properties": Dict[str, "SwaggerSchemaRef"]}, total=False)
+    SwaggerSchemaRef = Union[SwaggerSchema, SwaggerRef]
+    SwaggerContent = TypedDict("SwaggerContent", {"schema": SwaggerSchemaRef})
+    SwaggerResponse = TypedDict("SwaggerResponse", {"summary": SwaggerSummary, "content": Dict[str, SwaggerContent]})
+    SwaggerParameter = TypedDict("SwaggerParameter", {"name": str, "in": str, "required": bool, "type": str})
+    SwaggerSecurity = TypedDict("SwaggerSecurity", {"type": str, "name": str, "in": str})
+    SwaggerPath = TypedDict("SwaggerPath", {"responses": Dict[str, SwaggerResponse],
+                                            "parameters": List[SwaggerParameter],
+                                            "summary": SwaggerSummary,
+                                            "tags": List[str]}, total=False)
+    SwaggerType = TypedDict("SwaggerType", {"info": SwaggerInfo, "host": Optional[str], "schemes": List[str],
+                                            "basePath": str, "swagger": Optional[str], "openapi": Optional[str],
+                                            "tags": List[SwaggerTag], "paths": Dict[str, SwaggerPath],
+                                            "externalDocs": Optional[SwaggerDoc],
+                                            "securityDefinitions": Optional[Dict[str, SwaggerSecurity]]}, total=False)
 
 
 class CorniceSwaggerException(Exception):
@@ -44,7 +112,8 @@ class DefinitionHandler(object):
         :rtype: dict
         :returns: Swagger schema.
         """
-        return self._ref_recursive(self.type_converter(schema_node), self.ref, base_name)
+        name = self._get_schema_name(schema_node, base_name)
+        return self._ref_recursive(self.type_converter(schema_node), self.ref, name)
 
     def _ref_recursive(self, schema, depth, base_name=None):
         """
@@ -55,7 +124,7 @@ class DefinitionHandler(object):
             Base swagger schema.
         :param depth:
             How many levels of the swagger object schemas should be split into
-            swaggger definitions with JSON pointers. Default (0) is no split.
+            swagger definitions with JSON pointers. Default (0) is no split.
             You may use negative values to split everything.
         :param base_name:
             If schema doesn't have a name, the caller may provide it to be
@@ -70,18 +139,71 @@ class DefinitionHandler(object):
         if depth == 0:
             return schema
 
-        if schema["type"] != "object":
+        for kw in ['oneOf', 'allOf', 'anyOf', 'not']:
+            items = schema.get(kw)
+            if isinstance(items, list):
+                base_name = schema.get("title") or base_name
+                return self._process_items(schema, kw, items, depth, base_name)
+
+        if schema['type'] == 'array' and isinstance(schema['items'], dict):
+            ref_pointer = self._schema_object_to_pointer(schema, depth, base_name)
+            item_name = schema['items'].get('title')
+            if not item_name:
+                item_name = schema.get('title') + 'Item' or base_name + 'Item'
+            schema['items'] = self._ref_recursive(schema['items'], depth-1, item_name)
+            return ref_pointer
+
+        if schema['type'] != 'object':
             return schema
 
-        name = base_name or schema["title"]
+        return self._schema_object_to_pointer(schema, depth, base_name)
+
+    def _process_items(self, schema, list_type, item_list, depth, base_name):
+        ref_pointer = self._schema_object_to_pointer(schema, depth, base_name)
+        ref_list = []
+        for i, item in enumerate(item_list):
+            # recursively process nested keywords (eg: oneOf[allOf[x,y], allOf[x,z]])
+            for kw in ['oneOf', 'allOf', 'anyOf', 'not']:
+                sub_items = item.get(kw)
+                if isinstance(sub_items, list):
+                    sub_name = item.get('title') or (base_name + kw[0].upper() + kw[1:] + "Item" + str(i))
+                    ref = self._process_items(item, kw, sub_items, depth-1, sub_name)
+                    ref_list.append(ref)
+                    break
+            # otherwise process the item normally
+            else:
+                sub_name = item.get('title') or (base_name + "Item" + str(i))
+                ref = self._ref_recursive(item, depth-1, sub_name)
+                ref_list.append(ref)
+        self.definition_registry[base_name] = {list_type: ref_list}
+        return ref_pointer
+
+    @staticmethod
+    def _get_schema_name(schema, base_name):
+        """Returns the name or title of a schema node using
+        various getters if ``base_name`` is not defined."""
+        if base_name:
+            return base_name
+        # verify against valid strings to avoid retrieving a sub-schema
+        # named 'title' or 'name' within a mapping schema node.
+        for key in ['title', 'name']:
+            name = getattr(schema, key, None)
+            if isinstance(name, six.string_types) and len(name):
+                return name
+            name = schema.get(key)
+            if isinstance(name, six.string_types) and len(name):
+                return name
+        return type(schema).__name__
+
+    def _schema_object_to_pointer(self, schema, depth, base_name):
+        name = self._get_schema_name(schema, base_name)
 
         pointer = self.json_pointer + name
         for child_name, child in schema.get("properties", {}).items():
             schema["properties"][child_name] = self._ref_recursive(child, depth - 1)
 
         self.definition_registry[name] = schema
-
-        return {"$ref": pointer}
+        return {'$ref': pointer}
 
 
 class ParameterHandler(object):
@@ -129,12 +251,20 @@ class ParameterHandler(object):
             location = param_schema.name
             if location == "body":
                 name = param_schema.__class__.__name__
-                if name == "body":
-                    name = schema_node.__class__.__name__ + "Body"
+                if name == 'body':
+                    name = schema_node.__class__.__name__ + 'Body'
                 param = self.parameter_converter(location, param_schema)
-                param["name"] = name
+                param['name'] = name
                 if self.ref:
                     param = self._ref(param)
+                    # if definitions reference depth was also requested,
+                    # loop down body elements and replace the parameter reference
+                    if self.definitions.ref:
+                        body_ref = self.definitions.from_schema(param_schema, name)
+                        body_schema = self.definitions.definition_registry.get(name)
+                        param_ref = self.parameter_registry.get(name)
+                        param_ref["schema"] = body_ref
+
                 params.append(param)
 
             elif location in (("path", "header", "headers", "querystring", "GET")):
@@ -225,7 +355,7 @@ class ResponseHandler(object):
         for status, response_schema in schema_mapping.items():
             response = {}
             if response_schema.description:
-                response["description"] = response_schema.description
+                response['summary'] = response_schema.description
             else:
                 raise CorniceSwaggerException("Responses must have a description.")
 
@@ -252,6 +382,15 @@ class ResponseHandler(object):
             pointer = response_schema.__class__.__name__
             if self.ref:
                 response = self._ref(response, pointer)
+
+            # warning:
+            #   if using swagger 2.0, cannot have both examples and $ref, examples will be ignored
+            #   provide them here for OpenAPIv3 that can show both schemas and examples, or Swagger 2.0 without $ref
+            # see (at bottom): https://swagger.io/docs/specification/2-0/adding-examples/
+            examples = getattr(response_schema, 'examples', None)
+            if isinstance(examples, dict):
+                response['examples'] = examples
+
             responses[status] = response
 
         return responses
@@ -354,9 +493,34 @@ class CorniceSwagger(object):
     base_path = "/"
     """Base path of the documented API. Default is "/"."""
 
-    swagger = {"info": {}}
+    swagger = {'info': {}}  # type: SwaggerType
     """Base OpenAPI document that should be merged with the extracted info
     from the generate call."""
+
+    openapi_spec = 3
+    """OpenAPI specification version that should be used for generation. 
+    OpenAPI 3 is required  to support 'oneOf', 'allOf', 'anyOf' and 'not' keywords
+    as well as other multiple views of same method for different content-type."""
+
+    default_content_type = 'application/json'
+    """Default Content-Type to use for request/response body schemas. This is a 
+    required field for OpenAPI 3 specification generation, and it will be used if 
+    the views did not specify some value in another manner such as using the 
+    accept, renderer or content_type keywords. 
+    """
+
+    openapi_spec = 3
+    """OpenAPI specification version that should be used for generation. 
+    OpenAPI 3 is required  to support 'oneOf', 'allOf', 'anyOf' and 'not' keywords
+    as well as other multiple views of same method for different content-type."""
+
+    default_content_type = 'application/json'
+    """Default Content-Type to use for request/response body schemas. This is a 
+    required field for OpenAPI 3 specification generation, and it will be used if 
+    the views did not specify some value in another manner such as using the 
+    accept, renderer or content_type keywords. 
+    """
+
 
     def __init__(
         self,
@@ -408,10 +572,9 @@ class CorniceSwagger(object):
             self.definitions, ref=resp_ref, type_converter=type_converter
         )
 
-    def generate(
-        self, title=None, version=None, base_path=None, info=None, swagger=None, **kwargs
-    ):
-        """Generate a Swagger 2.0 documentation. Keyword arguments may be used
+    def generate(self, title=None, version=None, base_path=None,
+                 info=None, swagger=None, openapi_spec=2, **kwargs):
+        """Generate an OpenAPI documentation. Keyword arguments may be used
         to provide additional information to build methods as such ignores.
 
         :param title:
@@ -424,10 +587,19 @@ class CorniceSwagger(object):
             Swagger info field.
         :param swagger:
             Extra fields that should be provided on the swagger documentation.
+        :param openapi_spec:
+            The OpenAPI specification version to use for swagger documentation format.
 
         :rtype: dict
         :returns: Full OpenAPI/Swagger compliant specification for the application.
         """
+        openapi_spec = LooseVersion(str(openapi_spec)).version[0]
+        if openapi_spec not in [2, 3]:
+            raise CorniceSwaggerException('invalid OpenAPI specification version')
+        self.openapi_spec = openapi_spec
+        setattr(self.type_converter, "openapi_spec", self.openapi_spec)
+        setattr(self.parameter_converter, "openapi_spec", self.openapi_spec)
+
         title = title or self.api_title
         version = version or self.api_version
         info = info or self.swagger.get("info", {})
@@ -436,7 +608,13 @@ class CorniceSwagger(object):
 
         swagger = swagger.copy()
         info.update(title=title, version=version)
-        swagger.update(swagger="2.0", info=info, basePath=base_path)
+        swagger.update(info=info, basePath=base_path)
+        if self.openapi_spec == 2:
+            swagger.update(swagger='2.0')
+            swagger.pop("openapi", None)
+        else:
+            swagger.update(openapi='3.0.0')
+            swagger.pop("swagger", None)
 
         paths, tags = self._build_paths()
 
@@ -507,7 +685,9 @@ class CorniceSwagger(object):
             tags = self._get_tags(tags, service_tags)
 
             for method, view, args in service.definitions:
-                if method.lower() in map(str.lower, self.ignore_methods):
+                method_key = method.lower()
+
+                if method_key in map(str.lower, self.ignore_methods):
                     continue
 
                 op = self._extract_operation_from_view(view, args)
@@ -515,19 +695,22 @@ class CorniceSwagger(object):
                 if any(ctype in op.get("consumes", []) for ctype in self.ignore_ctypes):
                     continue
 
-                # XXX: Swagger doesn't support different schemas for for a same method
+                # XXX: Swagger 2.0 doesn't support different schemas for for a same method
                 # with different ctypes as cornice. If this happens, you may ignore one
                 # content-type from the documentation otherwise we raise an Exception
                 # Related to https://github.com/OAI/OpenAPI-Specification/issues/146
-                previous_definition = path_obj.get(method.lower())
-                if previous_definition:
-                    raise CorniceSwaggerException(
-                        (
-                            "Swagger doesn't support multiple "
-                            "views for a same method. You may "
-                            "ignore one."
-                        )
-                    )
+                previous_definition = path_obj.get(method_key)
+                if self.openapi_spec < 3:
+                    if previous_definition:
+                        raise CorniceSwaggerException(("Swagger doesn't support multiple "
+                                                       "views for a same method. You may "
+                                                       "ignore one."))
+                # move 'consumes' and 'produces' under their OpenAPI 3 'equivalents'
+                else:
+                    # OpenAPI 3 supports multiple schemas, but they should have different content-type
+                    op = self._convert_to_oas3(op)
+                    if previous_definition:
+                        self._validate_diff_oas3(op, previous_definition, path)
 
                 # If tag not defined and a default tag is provided
                 if "tags" not in op and self.default_tags:
@@ -563,10 +746,111 @@ class CorniceSwagger(object):
                 if not isinstance(op.get("security", []), list):
                     raise CorniceSwaggerException("security should be a list or callable")
 
-                path_obj[method.lower()] = op
+                if previous_definition:
+                    merge_dicts(path_obj[method_key], op)
+                else:
+                    path_obj[method_key] = op
             paths[path] = path_obj
 
         return paths, tags
+
+    def _convert_to_oas3(self, operations):
+        """
+        Extract 'consumes' and 'produces' objects from OpenAPI 2 (swagger)
+        definition and generates their equivalent for OpenAPI 3 format.
+
+        :param operations:
+            Pre-parsed specification of one operation.
+            (ie: route/method and optionally content-type combination)
+
+        :rtype: dict
+        :returns: Updated path definition.
+        """
+
+        consumes = operations.pop('consumes', [])
+        parameters = operations.get('parameters', {})
+        bodies = []
+        for param in parameters:
+            param_def = param
+            param_ref = None
+            if '$ref' in param:
+                param_ref = param
+                param_key = param['$ref'].split('/parameters/')[-1]
+                param_def = self.parameters.parameter_registry.get(param_key)
+            if param_def.get('in') == 'body':
+                parameters.remove(param)
+                bodies.append((param_def, param_ref))
+        if bodies:
+            # use-cases:
+            # (1) single view specified content_type=(<multiple-types>,...) but
+            #     only 1 body schema exists, suppose that view knows how to
+            #     parse every content-types structured with corresponding schema
+            # (2) single view specified content_type argument or content-type is given
+            #     as 'content_type="<ctype>"' argument to 'body = MappingSchema()' definition
+            # (3) many views each define their specific content_type as in (2),
+            #     results will be merged at the end
+            required = False
+            operations['requestBody'] = {'content': {}}
+            for body, body_ref in bodies:
+                body.pop('in')
+                if not len(consumes):
+                    consumes = [getattr(body, 'content_type', self.default_content_type)]
+                for ctype in consumes:
+                    required = required or body.pop('required', False)
+                    body_def = body_ref if body_ref else body
+                    operations['requestBody']['content'].update({ctype: body_def.copy()})
+            operations['requestBody']['required'] = required
+
+        produces = operations.pop('produces', [])
+        if not len(produces):
+            produces = [self.default_content_type]
+        for code in operations.get('responses', []):
+            if code == 'default':
+                continue
+            resp = operations['responses'][code]
+            if '$ref' in resp:
+                body = resp
+            else:
+                body = resp.pop('schema', None)
+            if body:
+                for ctype in produces:
+                    content = {'schema': body.copy()}
+                    examples = resp.pop('examples', None)
+                    if examples:
+                        content['examples'] = examples
+                    resp['content'] = {ctype: content}
+                    resp.pop('$ref', None)
+        return operations
+
+    def _validate_diff_oas3(self, operation, previous_definition, path):
+        """
+        Validates that two corresponding path operation definitions
+        (generated from two different views) each describe their respective
+        content-type without conflicts.
+        """
+        def _extract_info(op):
+
+            # TODO:
+            #  maybe keep/remove? to test
+            #  body can be duplicated because of different 'accept' views
+            #  response *should* not be duplicated because 1 view == 1 ctypes accept/renderer...
+            variations = []
+            #for ctype in op.get('requestBody', {}).get('content', []):
+            #    variations.append({"type": "requestBody", "ctype": ctype})
+            for code in op.get('responses', []):
+                contents = op['responses'][code].get('content', [])
+                for ctype in contents:
+                    variations.append({"type": "response", "ctype": ctype, "code": code})
+            return variations
+
+        var_oper = _extract_info(operation)
+        var_prev = _extract_info(previous_definition)
+        for vo in var_oper:
+            for vp in var_prev:
+                if vo == vp:
+                    raise CorniceSwaggerException(
+                        "Invalid path '{}' definitions specifies conflicting "
+                        "parameters {} at more than one place.".format(path, vo))
 
     def _extract_path_from_service(self, service):
         """
@@ -624,18 +908,23 @@ class CorniceSwagger(object):
             "responses": {"default": {"description": "UNDOCUMENTED RESPONSE"}},
         }
 
-        # If 'produces' are not defined in the view, try get from renderers
-        renderer = args.get("renderer", "")
+        # If 'produces' are not defined in the view, try get from accept/renderer arguments
+        renderer = args.get('accept', args.get('renderer', ''))
+        # if the input was literal content type, use it as is, otherwise infer from renderer method
+        # method should never have a '/' in it while content-type should always do
+
 
         is_json_renderer = (
             "json" in renderer  # allows for "json" or "simplejson"
             or renderer == Service.renderer  # default renderer is json.
         )
 
-        if is_json_renderer:
-            produces = ["application/json"]
-        elif renderer == "xml":
-            produces = ["text/xml"]
+        if '/' in renderer:
+            produces = [renderer]
+        elif is_json_renderer:  # allows for "json" or "simplejson"
+            produces = ['application/json']
+        elif 'xml' in renderer:  # all variants of 'xml' content-types
+            produces = ['text/xml']
         else:
             produces = None
 
@@ -656,26 +945,43 @@ class CorniceSwagger(object):
 
         # Get parameters from view schema
         is_colander = self._is_colander_schema(args)
+        deprecated = False
         if is_colander:
             schema = self._extract_transform_colander_schema(args)
             parameters = self.parameters.from_schema(schema)
+            deprecated = getattr(schema, 'deprecated', False)
         else:
             # Bail out for now
             parameters = None
         if parameters:
-            op["parameters"] = parameters
+            op['parameters'] = parameters
+
+        if not isinstance(deprecated, bool):
+            deprecated = False
+        if not deprecated and not isinstance(view, six.string_types):
+            deprecated = getattr(view, 'deprecated', False)
+        if deprecated is True:
+            op['deprecated'] = True
+
+        if not isinstance(deprecated, bool):
+            deprecated = False
+        if not deprecated and not isinstance(view, six.string_types):
+            deprecated = getattr(view, 'deprecated', False)
+        if deprecated is True:
+            op['deprecated'] = True
 
         # Get summary from docstring
-        if isinstance(view, str):
-            if "klass" in args:
-                ob = args["klass"]
-                view_ = getattr(ob, view.lower())
-                docstring = trim(view_.__doc__)
-        else:
-            docstring = str(trim(view.__doc__))
-
-        if docstring and self.summary_docstrings:
-            op["summary"] = docstring
+        if self.summary_docstrings:
+            docstring = None
+            if isinstance(view, six.string_types):
+                if 'klass' in args:
+                    ob = args['klass']
+                    view_ = getattr(ob, view.lower())
+                    docstring = trim(view_.__doc__)
+            else:
+                docstring = str(trim(view.__doc__))
+            if docstring:
+                op['summary'] = docstring
 
         # Get response definitions
         if "response_schemas" in args:
@@ -695,11 +1001,12 @@ class CorniceSwagger(object):
 
         return op
 
-    def _is_colander_schema(self, args):
-        schema = args.get("schema")
-        return isinstance(schema, colander.Schema) or (
-            inspect.isclass(schema) and issubclass(schema, colander.MappingSchema)
-        )
+    @staticmethod
+    def _is_colander_schema(args):
+        schema = args.get('schema')
+        return (isinstance(schema, colander.Schema) or
+                (inspect.isclass(schema)
+                and issubclass(schema, colander.MappingSchema)))
 
     def _extract_transform_colander_schema(self, args):
         """